requirements.txt
tensorflow-gpu
praw